Fix Linux Build Issues, Prep for Linux Implementation
